Handle images on page laod better.
It will scroll to the image on page load.
If you don't have it set to always use the big image, then the first time you press the resize hotkey, it will load the larger image and scroll to it.
The `fit-width` class is enabled on the image by default (and it sucks for tall images).